requirements.txt
langchain_community==0.4.1
langchain_core==1.0.4
langchain_openai==1.0.2
langchain_postgres==0.0.16
langchain_text_splitters==1.0.0
pydantic==2.12.4
python-dotenv==1.2.1
